+++ iohub/utils.py
@@ -99,45 +99,7 @@ def load_bg(bg_path, height, width, ROI=None):
     return bg_img_arr  # CYX
 
 
-def create_grid_from_coordinates(xy_coords, rows, columns):
-    """
-    Function to create a grid from XY-position coordinates.  Useful for generating HCS Zarr metadata.
 
-    Parameters
-    ----------
-    xy_coords:          (list) XY Stage position list in the order in which it was acquired: (X, Y) tuple.
-    rows:               (int) number of rows in the grid-like acquisition
-    columns:            (int) number of columns in the grid-like acquisition
-
-    Returns
-    -------
-    pos_index_grid      (array) A grid-like array mimicking the shape of the acquisition where the value in the array
-                                corresponds to the position index at that location.
-    """
-
-    coords = dict()
-    coords_list = []
-    for idx, pos in enumerate(xy_coords):
-        coords[idx] = pos
-        coords_list.append(pos)
-
-    # sort by X and then by Y
-    coords_list.sort(key=lambda x: x[0])
-    coords_list.sort(key=lambda x: x[1])
-
-    # reshape XY coordinates into their proper 2D shape
-    grid = np.reshape(coords_list, (rows, columns, 2))
-    pos_index_grid = np.zeros((rows, columns), "uint16")
-    keys = list(coords.keys())
-    vals = list(coords.values())
-
-    for row in range(rows):
-        for col in range(columns):
-
-            # append position index (key) into a final grid by indexed into the coordinate map (values)
-            pos_index_grid[row, col] = keys[vals.index(list(grid[row, col]))]
-
-    return pos_index_grid
 
 
 class MockEmitter:
