+++ iohub/zarr_converter.py
@@ -4,9 +4,48 @@ import numpy as np
 import tifffile as tiff
 from waveorder.io.writer import WaveorderWriter
 from waveorder.io.reader import WaveorderReader
-from iohub.utils import create_grid_from_coordinates
 import copy
 import json
+
+
+def create_grid_from_coordinates(xy_coords, rows, columns):
+    """
+    Function to create a grid from XY-position coordinates.  Useful for generating HCS Zarr metadata.
+
+    Parameters
+    ----------
+    xy_coords:          (list) XY Stage position list in the order in which it was acquired: (X, Y) tuple.
+    rows:               (int) number of rows in the grid-like acquisition
+    columns:            (int) number of columns in the grid-like acquisition
+
+    Returns
+    -------
+    pos_index_grid      (array) A grid-like array mimicking the shape of the acquisition where the value in the array
+                                corresponds to the position index at that location.
+    """
+
+    coords = dict()
+    coords_list = []
+    for idx, pos in enumerate(xy_coords):
+        coords[idx] = pos
+        coords_list.append(pos)
+
+    # sort by X and then by Y
+    coords_list.sort(key=lambda x: x[0])
+    coords_list.sort(key=lambda x: x[1])
+
+    # reshape XY coordinates into their proper 2D shape
+    grid = np.reshape(coords_list, (rows, columns, 2))
+    pos_index_grid = np.zeros((rows, columns), "uint16")
+    keys = list(coords.keys())
+    vals = list(coords.values())
+
+    for row in range(rows):
+        for col in range(columns):
+            # append position index (key) into a final grid by indexed into the coordinate map (values)
+            pos_index_grid[row, col] = keys[vals.index(list(grid[row, col]))]
+
+    return pos_index_grid
 
 
 class ZarrConverter:
@@ -132,7 +171,6 @@ class ZarrConverter:
             not self.summary_metadata
             or "AxisOrder" not in self.summary_metadata.keys()
         ):
-
             self.p_dim = 0
             self.t_dim = 1
             self.c_dim = 2
@@ -185,7 +223,6 @@ class ZarrConverter:
         ]
 
     def _get_position_coords(self):
-
         row_max = 0
         col_max = 0
         coords_list = []
@@ -202,7 +239,6 @@ class ZarrConverter:
         return coords_list, row_max + 1, col_max + 1
 
     def _generate_hcs_metadata(self):
-
         position_list, rows, cols = self._get_position_coords()
 
         position_grid = create_grid_from_coordinates(position_list, rows, cols)
@@ -396,7 +432,6 @@ class ZarrConverter:
         chan_names = self._get_channel_names()
         self._get_position_names()
         for pos in range(self.p):
-
             clims = self.get_channel_clims(pos)
             name = self.pos_names[pos] if self.replace_position_names else None
             self.writer.init_array(
